+++ tests/setup-dom.ts
@@ -1,0 +1,4 @@
+import { configureDomAdapter } from "../src/dom";
+import { BROWSER_DOM_ADAPTER } from "../site/browser-dom-adapter";
+
+configureDomAdapter(BROWSER_DOM_ADAPTER);